[draft] Add support for raw pointers as CornerStorage
By using `Dune::resolvePtrOrRef()` we can support raw pointers
as `CornerStorage` in `MultiLinearGeometry` and thus have a
non-owning corner storage while still being default constructible.

The test adds a third test case with raw pointer as `CornerStorage`
and also checks default construction and assignment of `MultiLinearGeometry`
if the `CornerStorage` is default constructible.

Warning: This is a **DRAFT** implementation to demonstrate the approach.